v8.2.0: Release all sweep patients through no-reward controls
