--- .github/workflows/temporary-issue565-v8-2-0-executor.yml
@@ -1,118 +1,0 @@
-name: Temporary Issue 565 v8.2.0 Executor
-
-on:
-  pull_request:
-    types: [opened, reopened, synchronize]
-
-permissions:
-  contents: read
-  pull-requests: read
-
-concurrency:
-  group: temporary-issue565-v8-2-0
-  cancel-in-progress: false
-
-jobs:
-  execute:
-    if: >-
-      github.repository == 'Conroy1988/missionchief-toolkit-assets' &&
-      github.event.pull_request.number == 566 &&
-      github.event.pull_request.head.ref == 'feature/transport-sweep-no-reward-release-v8-2-0'
-    runs-on: ubuntu-latest
-    timeout-minutes: 35
-    steps:
-      - name: Require owner branch token
-        env:
-          OWNER_TOKEN: ${{ secrets.DEVELOPMENT_PR_TOKEN }}
-        shell: bash
-        run: |
-          set -euo pipefail
-          [[ -n "$OWNER_TOKEN" ]] || { echo 'DEVELOPMENT_PR_TOKEN is required.' >&2; exit 1; }
-
-      - name: Check out exact pull-request head
-        uses: actions/checkout@9c091bb21b7c1c1d1991bb908d89e4e9dddfe3e0 # v7
-        with:
-          ref: ${{ github.event.pull_request.head.sha }}
-          fetch-depth: 0
-          token: ${{ secrets.DEVELOPMENT_PR_TOKEN }}
-
-      - name: Apply, clean, validate and publish
-        env:
-          EXPECTED_BASE: bf4c8b23eda4cb7982df4c79228bce1e5ad3ea69
-          EXPECTED_PACKAGE_COMMIT: bc8c9bf1d97cc0bf9a6232d4674eaa4db11a1e26
-          EXPECTED_HEAD: ${{ github.event.pull_request.head.sha }}
-          BRANCH: feature/transport-sweep-no-reward-release-v8-2-0
-          MARKER_FIX: .github/development-packages/issue565_package_marker_fix.py
-          PACKAGE: .github/development-packages/issue565_transport_sweep_no_reward_v8_2_0.py
-          SCOPE_MARKER_FIX: .github/development-packages/issue565_scope_marker_fix.py
-          SCOPE_FIX: .github/development-packages/issue565_verified_vehicle_scope_fix.py
-          NATIVE_CONTRACT_FIX: .github/development-packages/issue565_native_contract_scope_fix.py
-          STATIC_ORDER_FIX: .github/development-packages/issue565_static_order_fix.py
-          RUNTIME_HARNESS_FIX: .github/development-packages/issue565_runtime_harness_fix.py
-          ALLIANCE_CONTRACT_FIX: .github/development-packages/issue565_release_forward_alliance_contract.py
-          FINALIZE: .github/development-packages/issue565_finalize_v8_2_0.py
-          EXECUTOR: .github/workflows/temporary-issue565-v8-2-0-executor.yml
-        shell: bash
-        run: |
-          set -euo pipefail
-          exec > >(tee "$RUNNER_TEMP/issue565-executor.log") 2>&1
-          echo stage=verify
-          git merge-base --is-ancestor "$EXPECTED_PACKAGE_COMMIT" HEAD
-          git fetch --no-tags origin main
-          test "$(git rev-parse origin/main)" = "$EXPECTED_BASE"
-          git merge-base --is-ancestor origin/main HEAD
-          test -f "$MARKER_FIX"
-          test -f "$PACKAGE"
-          test -f "$SCOPE_MARKER_FIX"
-          test -f "$SCOPE_FIX"
-          test -f "$NATIVE_CONTRACT_FIX"
-          test -f "$STATIC_ORDER_FIX"
-          test -f "$RUNTIME_HARNESS_FIX"
-          test -f "$ALLIANCE_CONTRACT_FIX"
-          test -f "$FINALIZE"
-          test -f "$EXECUTOR"
-
-          echo stage=apply
-          python3 "$MARKER_FIX"
-          python3 "$SCOPE_MARKER_FIX"
-          python3 "$PACKAGE"
-          python3 "$SCOPE_FIX"
-          python3 "$NATIVE_CONTRACT_FIX"
-          python3 "$STATIC_ORDER_FIX"
-          python3 "$RUNTIME_HARNESS_FIX"
-          python3 "$ALLIANCE_CONTRACT_FIX"
-          python3 "$FINALIZE"
-          rm -f "$MARKER_FIX" "$PACKAGE" "$SCOPE_MARKER_FIX" "$SCOPE_FIX" "$NATIVE_CONTRACT_FIX" "$STATIC_ORDER_FIX" "$RUNTIME_HARNESS_FIX" "$ALLIANCE_CONTRACT_FIX" "$FINALIZE" "$EXECUTOR"
-          find .github/development-packages -type d -empty -delete 2>/dev/null || true
-
-          echo stage=dependencies
-          npm install --no-save --package-lock=false --ignore-scripts jsdom@26.1.0
-
-          echo stage=validate
-          python3 .github/scripts/validate_userscript.py
-          node --check src/MissionChief_Map_Command_Toolkit.user.js
-          cmp --silent dist/MissionChief_Map_Command_Toolkit.user.js dist/MissionChief_Map_Command_Toolkit.txt
-          python3 .github/scripts/test_issue565_transport_sweep_no_reward.py
-          node .github/scripts/test_issue565_transport_sweep_no_reward_runtime.mjs
-          bash .github/scripts/run_userscript_preflight.sh
-
-          echo stage=clean-dependencies
-          rm -rf node_modules
-          test -z "$(git ls-files node_modules)"
-
-          echo stage=publish
-          git config user.name "Conroy1988"
-          git config user.email "27301455+Conroy1988@users.noreply.github.com"
-          git add -A
-          git diff --cached --quiet && { echo 'No candidate changes produced.' >&2; exit 1; }
-          git commit -m "v8.2.0: Release all sweep patients through no-reward controls"
-          git push --force-with-lease="refs/heads/${BRANCH}:${EXPECTED_HEAD}" origin "HEAD:${BRANCH}"
-
-      - name: Upload executor diagnostics
-        if: always()
-        uses: actions/upload-artifact@043fb46d1a93c77aae656e7c1c64a875d1fc6a0a # v7.0.1
-        with:
-          name: issue565-v8-2-0-executor-${{ github.sha }}
-          path: ${{ runner.temp }}/issue565-executor.log
-          if-no-files-found: warn
-          retention-days: 14
